Say what a branch is doing in GitHub's own icons
- A pull request is purple whether open or merged, yellow with the
  queue icon while it waits there, grey as a draft and red closed,
  and a repository's own branch is green rather than another grey
- Reviews say their verdict: a filled green tick approved, a
  filled red cross asked for changes, and a clock for one that is
  required but has not happened, which is waiting, not failing
- Unresolved conversations show a comment icon, counted from what
  the conversation pane has already read: GitHub answers that only
  through GraphQL, far too dear to ask per branch per poll
- The line now reads number, CI, reviews, conversations, then the
  commit counts, and the unread dot moved to the row's edge where
  a column of them can be scanned
- A push of rewritten history leases rather than being refused as
  a non-fast-forward, which is what amending a pushed branch left
