+++ README.md
@@ -47,6 +47,12 @@ before shipping.
   since each was last viewed, open pull requests, mergeability and
   uncommitted or unpushed work, and a worktree can be marked unread to
   revisit (so you always know where you are needed)
+- **Says** what a branch is doing in GitHub's own icons: purple for a pull
+  request, open or merged, yellow while it waits in the merge queue, grey
+  for a draft, green for a repository's own branch; then its number, a
+  CI dot, a review that has approved, asked for changes or not happened
+  yet, any unresolved conversations, and its commit counts last (so one
+  glance across the sidebar says where everything stands)
 - **Notifies** you when an agent finishes or its output stalls (so you never sit
   polling a terminal)
 - **Renders** terminals locally from a `tmux` control mode client, so selecting,
